+++ PKHeX.Core/Legality/Enums/CheckIdentifier.cs
@@ -1,0 +1,35 @@
+﻿namespace PKHeX.Core
+{
+    internal enum CheckIdentifier
+    {
+        Move,
+        RelearnMove,
+        Encounter,
+        History,
+        ECPID,
+        Shiny,
+        EC,
+        PID,
+        Gender,
+        EVs,
+        Language,
+        Nickname,
+        Trainer,
+        IVs,
+        None,
+        Level,
+        Ball,
+        Memory,
+        Geography,
+        Form,
+        Egg,
+        Misc,
+        Fateful,
+        Ribbon,
+        Training,
+        Ability,
+        Evolution,
+        Special,
+        Nature
+    }
+}
