+++ PKHeX.Core/Legality/Enums/Severity.cs
@@ -1,0 +1,11 @@
+﻿namespace PKHeX.Core
+{
+    public enum Severity
+    {
+        Indeterminate = -2,
+        Invalid = -1,
+        Fishy = 0,
+        Valid = 1,
+        NotImplemented = 2,
+    }
+}
